feat(activities): enhance activity schema and add utility functions
