+++ lib/db/operations/activities/mutations/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Activity Mutation Operations Exports
+ */
+export * from './mutations';
