+++ lib/db/operations/activities/queries/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Activity Query Operations Exports
+ */
+export * from './queries';
