+++ lib/db/operations/activities/stats/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Activity Statistics Operations Exports
+ */
+export * from './stats';
