+++ lib/db/operations/activities/types/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Activity Types Exports
+ */
+export type {
+  Activity,
+  NewActivity,
+  GetActivitiesOptions,
+  ActivityStats,
+  GetActivityStatsOptions,
+  ActivityFilterOptions,
+} from './types';
