+++ lib/db/operations/users/index.ts
@@ -1,0 +1,13 @@
+/**
+ * User Operations
+ * Provides Create, Read, Update, Delete operations for users
+ *
+ * This module is organized into:
+ * - queries.ts: Read operations (cached with React cache)
+ * - mutations.ts: Write operations (create, update, delete)
+ * - types.ts: Type definitions
+ */
+
+export * from './queries';
+export * from './mutations';
+export * from './types';
