+++ lib/db/operations/users/mutations/index.ts
@@ -1,0 +1,4 @@
+/**
+ * User Mutation Operations Exports
+ */
+export * from './mutations';
