+++ lib/db/operations/users/queries/index.ts
@@ -1,0 +1,4 @@
+/**
+ * User Query Operations Exports
+ */
+export * from './queries';
