+++ lib/db/operations/users/types/index.ts
@@ -1,0 +1,4 @@
+/**
+ * User Types Exports
+ */
+export * from './types';
