+++ lib/db/schema/activities.ts
@@ -19,54 +19,60 @@ import { users } from './users';
 export const activities = pgTable(
   'activities',
   {
+    // Core identification
     id: uuid('id').defaultRandom().primaryKey(),
     userId: uuid('user_id')
       .references(() => users.id, { onDelete: 'cascade' })
       .notNull(),
     stravaActivityId: varchar('strava_activity_id', { length: 255 }).unique(),
-    name: varchar('name', { length: 255 }).notNull(),
-    description: varchar('description', { length: 1000 }),
-    type: varchar('type', { length: 50 }).notNull(), // e.g., 'Ride', 'Run', 'Walk'
-    sportType: varchar('sport_type', { length: 50 }), // e.g., 'MountainBikeRide', 'GravelRide'
 
-    // Distance and duration
+    // Basic information (always required)
+    name: varchar('name', { length: 255 }).notNull(),
+    type: varchar('type', { length: 50 }).notNull(), // e.g., 'Ride', 'Run', 'Walk'
+    startDate: timestamp('start_date').notNull(),
+
+    // Optional metadata
+    description: varchar('description', { length: 1000 }),
+    sportType: varchar('sport_type', { length: 50 }), // More specific: 'MountainBikeRide', 'GravelRide'
+    timezone: varchar('timezone', { length: 100 }),
+
+    // Core metrics (commonly available)
     distance: real('distance'), // in meters
     movingTime: integer('moving_time'), // in seconds
     elapsedTime: integer('elapsed_time'), // in seconds
-
-    // Elevation
     totalElevationGain: real('total_elevation_gain'), // in meters
-    elevHigh: real('elev_high'), // in meters
-    elevLow: real('elev_low'), // in meters
-
-    // Performance metrics
     averageSpeed: real('average_speed'), // in meters per second
-    maxSpeed: real('max_speed'), // in meters per second
-    averageCadence: real('average_cadence'),
-    averageHeartrate: real('average_heartrate'),
-    maxHeartrate: real('max_heartrate'),
-    calories: real('calories'),
 
-    // Location
-    startLatlng: jsonb('start_latlng').$type<[number, number]>(),
-    endLatlng: jsonb('end_latlng').$type<[number, number]>(),
-    locationCity: varchar('location_city', { length: 100 }),
+    // Extended elevation data (less commonly used)
+    elevHigh: real('elev_high'), // in meters - maximum elevation point
+    elevLow: real('elev_low'), // in meters - minimum elevation point
+
+    // Performance metrics (requires sensors - often NULL)
+    maxSpeed: real('max_speed'), // in meters per second
+    averageCadence: real('average_cadence'), // RPM - requires cadence sensor
+    averageHeartrate: real('average_heartrate'), // BPM - requires heart rate monitor
+    maxHeartrate: real('max_heartrate'), // BPM - requires heart rate monitor
+    calories: real('calories'), // Calculated/estimated
+
+    // Location data
+    startLatlng: jsonb('start_latlng').$type<[number, number]>(), // [latitude, longitude]
+    endLatlng: jsonb('end_latlng').$type<[number, number]>(), // Often NULL, can be derived
+    locationCity: varchar('location_city', { length: 100 }), // Can be derived from coordinates
     locationState: varchar('location_state', { length: 100 }),
     locationCountry: varchar('location_country', { length: 100 }),
 
-    // Activity metadata
-    startDate: timestamp('start_date').notNull(),
-    timezone: varchar('timezone', { length: 100 }),
-    isManual: boolean('is_manual').default(false),
-    isPrivate: boolean('is_private').default(false),
+    // Activity flags
+    isManual: boolean('is_manual').default(false).notNull(),
+    isPrivate: boolean('is_private').default(false).notNull(),
 
-    // Map and summary
-    mapPolyline: varchar('map_polyline', { length: 10000 }),
-    mapSummaryPolyline: varchar('map_summary_polyline', { length: 2000 }),
+    // Map data (large - 2-10KB per activity, consider lazy loading)
+    mapPolyline: varchar('map_polyline', { length: 10000 }), // Full resolution route
+    mapSummaryPolyline: varchar('map_summary_polyline', { length: 2000 }), // Simplified route
 
-    // Additional data
+    // Flexible storage for API-specific or future data
     metadata: jsonb('metadata').$type<Record<string, any>>(),
 
+    // Audit timestamps
     createdAt: timestamp('created_at').defaultNow().notNull(),
     updatedAt: timestamp('updated_at').defaultNow().notNull(),
   },
