+++ lib/db/schema/index.ts
@@ -5,3 +5,6 @@
 
 export * from './users';
 export * from './activities';
+
+// Query helpers for selective field loading
+export * from './helpers/activityFields';
